iot-backend/requirements.txt
Flask==2.1.2
requests==2.25.1
paho-mqtt==1.6.1
